% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_functions.R
\name{add_new_patches}
\alias{add_new_patches}
\title{Add new patches to meet conservation targets}
\usage{
add_new_patches(minpatch_data, verbose = TRUE)
}
\arguments{
\item{minpatch_data}{List containing all MinPatch data structures (including prioritizr objects)}

\item{verbose}{Logical, whether to print progress}
}
\value{
Updated minpatch_data with new patches added
}
\description{
Stage 2 of MinPatch: Add new patches using BestPatch algorithm
}
\keyword{internal}
