% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_functions.R
\name{add_patch_centered_on_unit}
\alias{add_patch_centered_on_unit}
\title{Add patch centered on specified planning unit}
\usage{
add_patch_centered_on_unit(minpatch_data, center_unit_id)
}
\arguments{
\item{minpatch_data}{List containing all MinPatch data structures}

\item{center_unit_id}{ID of the unit to center the patch on}
}
\value{
Updated unit_dict with new patch added
}
\description{
Add patch centered on specified planning unit
}
\keyword{internal}
