% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_functions.R
\name{calculate_feature_conservation}
\alias{calculate_feature_conservation}
\title{Calculate current feature conservation amounts}
\usage{
calculate_feature_conservation(minpatch_data)
}
\arguments{
\item{minpatch_data}{List containing all MinPatch data structures}
}
\value{
Named vector of conserved amounts for each feature
}
\description{
Calculates how much of each feature is currently conserved
}
\keyword{internal}
