% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_functions.R
\name{calculate_feature_representation}
\alias{calculate_feature_representation}
\title{Calculate feature representation in solution}
\usage{
calculate_feature_representation(minpatch_data)
}
\arguments{
\item{minpatch_data}{List containing all MinPatch data structures including prioritizr objects}
}
\value{
Data frame with feature representation statistics
}
\description{
Calculates how much of each conservation feature is represented
in the current solution using prioritizr functions where possible
}
