% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittling_functions.R
\name{can_remove_unit}
\alias{can_remove_unit}
\title{Check if a planning unit can be removed}
\usage{
can_remove_unit(unit_id, minpatch_data)
}
\arguments{
\item{unit_id}{ID of unit to potentially remove}

\item{minpatch_data}{List containing all MinPatch data structures}
}
\value{
Logical indicating if unit can be removed
}
\description{
Checks multiple criteria to determine if removing a unit is acceptable:
1. Doesn't violate conservation targets
2. Doesn't make any patch too small
3. Doesn't increase total cost
4. Doesn't split patches into non-viable pieces
}
\keyword{internal}
