% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structures.R
\name{extract_locked_out_constraints}
\alias{extract_locked_out_constraints}
\title{Extract locked-out constraint indices from prioritizr problem}
\usage{
extract_locked_out_constraints(prioritizr_problem, verbose = TRUE)
}
\arguments{
\item{prioritizr_problem}{A prioritizr problem object}

\item{verbose}{Logical, whether to print messages}
}
\value{
Integer vector of locked-out planning unit indices
}
\description{
Extract locked-out constraint indices from prioritizr problem
}
\keyword{internal}
