% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittling_functions.R
\name{find_edge_units}
\alias{find_edge_units}
\title{Find edge planning units}
\usage{
find_edge_units(minpatch_data)
}
\arguments{
\item{minpatch_data}{List containing all MinPatch data structures}
}
\value{
Character vector of edge unit IDs
}
\description{
Identifies planning units that are on the edge of selected areas
(have at least one unselected neighbor). Optimized with vector pre-allocation.
}
\keyword{internal}
