% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_functions.R
\name{identify_unmet_targets}
\alias{identify_unmet_targets}
\title{Identify features with unmet targets}
\usage{
identify_unmet_targets(minpatch_data)
}
\arguments{
\item{minpatch_data}{List containing all MinPatch data structures including prioritizr objects}
}
\value{
Character vector of feature IDs with unmet targets
}
\description{
Uses prioritizr functions to identify unmet targets from minpatch_data
}
\keyword{internal}
