% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minpatch.R
\docType{package}
\name{minpatch-package}
\alias{minpatch}
\alias{minpatch-package}
\title{MinPatch for R: Post-processing prioritizr solutions to ensure minimum patch sizes}
\description{
This package provides functions to post-process conservation planning solutions
from prioritizr to ensure all protected areas meet user-defined minimum size
thresholds, following the methodology described in Smith et al. (2010).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SpatialPlanning/minpatch}
  \item Report bugs at \url{https://github.com/SpatialPlanning/minpatch/issues}
}

}
\author{
\strong{Maintainer}: Jason D. Everett \email{DrJasonEverett@gmail.com} (\href{https://orcid.org/0000-0002-6681-8054}{ORCID})

Authors:
\itemize{
  \item Anthony J. Richardson \email{a.richardson1@uq.edu.au} (\href{https://orcid.org/0000-0002-9289-7366}{ORCID})
  \item Robert J. Smith (Original MinPatch algorithm author)
}

}
