% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_functions.R
\name{print_minpatch_summary}
\alias{print_minpatch_summary}
\title{Print MinPatch results summary}
\usage{
print_minpatch_summary(minpatch_result)
}
\arguments{
\item{minpatch_result}{Result object from run_minpatch function}
}
\description{
Prints a formatted summary of MinPatch processing results
}
