% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittling_functions.R
\name{removal_increases_cost}
\alias{removal_increases_cost}
\title{Check if removing unit would increase cost}
\usage{
removal_increases_cost(unit_id, minpatch_data)
}
\arguments{
\item{unit_id}{ID of unit to potentially remove}

\item{minpatch_data}{List containing all MinPatch data structures}
}
\value{
Logical indicating if removal would increase cost
}
\description{
Check if removing unit would increase cost
}
\keyword{internal}
