% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittling_functions.R
\name{removal_makes_patch_too_small}
\alias{removal_makes_patch_too_small}
\title{Check if removing unit would make its patch too small}
\usage{
removal_makes_patch_too_small(unit_id, minpatch_data)
}
\arguments{
\item{unit_id}{ID of unit to potentially remove}

\item{minpatch_data}{List containing all MinPatch data structures}
}
\value{
Logical indicating if removal would make patch too small
}
\description{
Check if removing unit would make its patch too small
}
\keyword{internal}
