% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittling_functions.R
\name{removal_splits_patch_nonviably}
\alias{removal_splits_patch_nonviably}
\title{Check if removing unit would split patch into non-viable pieces}
\usage{
removal_splits_patch_nonviably(unit_id, minpatch_data)
}
\arguments{
\item{unit_id}{ID of unit to potentially remove}

\item{minpatch_data}{List containing all MinPatch data structures}
}
\value{
Logical indicating if removal would create non-viable patches
}
\description{
Check if removing unit would split patch into non-viable pieces
}
\keyword{internal}
