% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittling_functions.R
\name{removal_violates_targets}
\alias{removal_violates_targets}
\title{Check if removing unit would violate conservation targets}
\usage{
removal_violates_targets(unit_id, minpatch_data)
}
\arguments{
\item{unit_id}{ID of unit to potentially remove}

\item{minpatch_data}{List containing all MinPatch data structures}
}
\value{
Logical indicating if removal would violate targets
}
\description{
Check if removing unit would violate conservation targets
}
\keyword{internal}
