% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_functions.R
\name{remove_small_patches_from_solution}
\alias{remove_small_patches_from_solution}
\title{Remove small patches from solution}
\usage{
remove_small_patches_from_solution(minpatch_data, patch_dict, min_patch_size)
}
\arguments{
\item{minpatch_data}{List containing all MinPatch data structures}

\item{patch_dict}{Named list of patches}

\item{min_patch_size}{Minimum patch size threshold}
}
\value{
Updated minpatch_data with small patches removed
}
\description{
Stage 1 of MinPatch: Remove patches smaller than minimum size threshold
}
\keyword{internal}
