% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittling_functions.R
\name{simulated_whittling}
\alias{simulated_whittling}
\title{Simulated whittling to remove unnecessary planning units}
\usage{
simulated_whittling(minpatch_data, verbose = TRUE)
}
\arguments{
\item{minpatch_data}{List containing all MinPatch data structures (including prioritizr objects)}

\item{verbose}{Logical, whether to print progress}
}
\value{
Updated minpatch_data with unnecessary units removed
}
\description{
Stage 3 of MinPatch: Remove planning units that are not needed to meet targets,
reduce fragmentation, or meet minimum patch size requirements
}
\keyword{internal}
